--- src/Protocol.hs
@@ -1,35 +1,0 @@
-{-# LANGUAGE DeriveGeneric #-}
-
-module Protocol (
-  Response(..),
-  Request(..),
-  emptyRequest,
-  emptyResponse) where
-
-import           Protocol.Expectation
-import           Protocol.Test
-import           GHC.Generics
-import           Data.Aeson
-
-data Request = Request {
-  content  :: String,
-  test     :: String,
-  extra    :: String,
-  expectations ::  [Expectation]
-} deriving (Show, Generic)
-
-data Response =  Response {
-  exit  :: String,
-  out   :: String,
-  testResults :: [TestResult],
-  expectationResults :: [ExpectationResult]
-} deriving (Eq, Show, Generic)
-
-instance FromJSON Request
-instance ToJSON Response
-
-emptyRequest :: Request
-emptyRequest = Request "" "" "" []
-
-emptyResponse :: Response
-emptyResponse = Response "" "" [] []
