+++ src/Protocol/Test.hs
@@ -1,15 +1,35 @@
 {-# LANGUAGE DeriveGeneric #-}
 
 module Protocol.Test (
-  TestResult(..)) where
+  Response(..),
+  Request(..),
+  emptyRequest,
+  emptyResponse) where
 
+import           Protocol.Test.Expectation
+import           Protocol.Test.Test
 import           GHC.Generics
 import           Data.Aeson
 
-data TestResult = TestResult {
-  title  :: String,
-  status :: String,
-  result :: String
-} deriving (Show, Eq, Generic)
+data Request = Request {
+  content  :: String,
+  test     :: String,
+  extra    :: String,
+  expectations ::  [Expectation]
+} deriving (Show, Generic)
 
-instance ToJSON TestResult
+data Response =  Response {
+  exit  :: String,
+  out   :: String,
+  testResults :: [TestResult],
+  expectationResults :: [ExpectationResult]
+} deriving (Eq, Show, Generic)
+
+instance FromJSON Request
+instance ToJSON Response
+
+emptyRequest :: Request
+emptyRequest = Request "" "" "" []
+
+emptyResponse :: Response
+emptyResponse = Response "" "" [] []
